[bugfix] Select the first element in DataComboBox if required
We had an issue today where we tried to force transition to the first
element of the transition list to trigger an abort behaviour on an FSM.
The first element appear selected by default, but in that case clicking
"Force transition" didn't work as the DataComboBox item didn't consider
the element as selected. Furthermore, clicking on it explicitely would
not update it's state, as it is the same element as the currently
selected one.

This commit fixes the behaviour by ensuring that the first element is
selected if it is a required element, and selecting element of index -1
by default when not required.