fix(panel/schema): Fix handling of nested object required properties
When schemas contain nested objects, the child's required properties
should only be required when the parent object is itself required.

e.g
Schema1:
  prop1*
  ChildSchema:
    child1*

child1* is only required if Schema1 is required (* represents a required
property)

This was already handled when including a schema through $ref, but not
when nesting objects directly.