fix(panel/schema): dependent combo box work after reset
This annoying bug hapenned in the following scenario:
- Add a task from the gui that has QComboBox depending on gui static
data (robot names, robot frames, surfaces, etc). For example the
transform task has combox for robot names and frames, with the frames
dependent on the robot name.
- Add the task to the solver
- This resets the GUI for the task
- Try selecting a robot/frame, it does not work

I am not sure how Qt handles disconnect/reconnect of signals, but
apparently the previous implementation is not how it should be done.